+++ R/SME.R
@@ -177,6 +177,12 @@
 #'     identification). Ties (equal score) are represented by using the same
 #'     rank, defaults to `"1"` if there is no ranking system used.
 #'
+#' @param opt_identifier `character` with the identifier to be used for optional
+#'     columns passed through `...`. This is used to create the column names for
+#'     optional columns in the format `opt_{opt_identifier}_{column_name}`. If
+#'     not provided, the default is `"global"`. Must be of length 1 or matching
+#'     the number of opt columns provided in `...`.
+#'
 #' @param mtd two-column `matrix` or `data.frame` with the metadata (MTD)
 #'     definition of the data set. The first column needs to contain the
 #'     metadata field names, the second the corresponding values. See
@@ -266,7 +272,7 @@ smeCreate <- function(..., evidence_input_id = character(),
                        identification_method = character(),
                        ms_level = character(),
                        id_confidence_measure = NULL,
-                       rank = 1, mtd = NULL) {
+                       rank = 1, opt_identifier = "global", mtd = NULL) {
     if (!length(evidence_input_id))
         stop("The argument 'evidence_input_id' is mandatory.")
     if (!length(exp_mass_to_charge))
@@ -312,7 +318,7 @@ smeCreate <- function(..., evidence_input_id = character(),
         x <- cbind(x, smeIdConfidenceMeasure(id_confidence_measure, mtd, l))
 
     ## Add optional columns
-    x <- .add_opt_cols(x = x, ...)
+    x <- .add_opt_cols(x = x, identifier = opt_identifier, ...)
     smeSort(x)
 }